requirements-dev.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
decorators==2.0.7
Django==5.2.13
django-environ==0.12.0
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
idna==3.11
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
phonenumbers==9.0.28
psycopg2-binary==2.9.10
PyJWT==2.12.1
pyTelegramBotAPI==4.25.0
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
sqlparse==0.5.5
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3